requirements.txt
PyQt6
PyPDF2
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
